refactor: split shared identity logic by ownership
Move GAIE reference handling, GVK selection, resolution, objective validation,
and pool selector extraction into internal/gaie.

Move SPIRE Controller Manager ClusterSPIFFEID labels, naming, rendering,
sync, merge, and drift helpers into internal/spirecm.

Keep internal/identity focused on pure identity planning and update controller
and inspection call sites to compose the narrower packages without changing
behavior.
