+++ internal/controller/inferenceidentitybinding_controller.go
@@ -35,7 +35,8 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	kleymv1alpha1 "github.com/sonda-red/kleym/api/v1alpha1"
-	identitypkg "github.com/sonda-red/kleym/internal/identity"
+	"github.com/sonda-red/kleym/internal/gaie"
+	"github.com/sonda-red/kleym/internal/spirecm"
 )
 
 const (
@@ -82,9 +83,9 @@ const (
 )
 
 var (
-	inferenceObjectiveGVKs = identitypkg.InferenceObjectiveGVKs()
-	inferencePoolGVKs      = identitypkg.InferencePoolGVKs()
-	clusterSPIFFEIDGVK     = identitypkg.ClusterSPIFFEIDGVK()
+	inferenceObjectiveGVKs = gaie.InferenceObjectiveGVKs()
+	inferencePoolGVKs      = gaie.InferencePoolGVKs()
+	clusterSPIFFEIDGVK     = spirecm.ClusterSPIFFEIDGVK()
 )
 
 // InferenceIdentityBindingReconciler reconciles a InferenceIdentityBinding object
@@ -285,10 +286,11 @@ func (r *InferenceIdentityBindingReconciler) Reconcile(
 	}
 
 	primaryIdentity := desiredState.identities[0]
-	r.recordEventf(binding, corev1.EventTypeNormal, "Reconciled", "reconciled ClusterSPIFFEID %q", primaryIdentity.Name)
+	primaryClusterSPIFFEIDName := desiredClusterSPIFFEID(binding, primaryIdentity).GetName()
+	r.recordEventf(binding, corev1.EventTypeNormal, "Reconciled", "reconciled ClusterSPIFFEID %q", primaryClusterSPIFFEIDName)
 	logger.Info(
 		"reconciled successfully",
-		logKeyClusterSPIFFEID, primaryIdentity.Name,
+		logKeyClusterSPIFFEID, primaryClusterSPIFFEIDName,
 		logKeySpiffeID, primaryIdentity.SpiffeID,
 	)
 
@@ -437,7 +439,7 @@ func (r *InferenceIdentityBindingReconciler) computeDesiredState(
 	logger := logf.FromContext(ctx)
 
 	mode := effectiveMode(binding.Spec.Mode)
-	poolRef, err := identitypkg.BindingPoolRef(binding)
+	poolRef, err := gaie.BindingPoolRef(binding)
 	if err != nil {
 		return desiredBindingState{}, newStateError(conditionTypeInvalidRef, "InvalidPoolRef", err.Error())
 	}
@@ -458,7 +460,7 @@ func (r *InferenceIdentityBindingReconciler) computeDesiredState(
 	)
 
 	var objective *unstructured.Unstructured
-	objectiveRef, hasObjectiveRef, err := identitypkg.BindingObjectiveRef(binding)
+	objectiveRef, hasObjectiveRef, err := gaie.BindingObjectiveRef(binding)
 	if err != nil {
 		return desiredBindingState{}, newStateError(conditionTypeInvalidRef, "InvalidObjectiveRef", err.Error())
 	}
@@ -480,7 +482,7 @@ func (r *InferenceIdentityBindingReconciler) computeDesiredState(
 			logKeyObjective, namespacedBindingKey(objective.GetNamespace(), objective.GetName()),
 			logKeyObjectiveGVK, objective.GroupVersionKind().String(),
 		)
-		if err := identitypkg.ValidateObjectiveTargetsPool(objective, pool, binding.Namespace); err != nil {
+		if err := gaie.ValidateObjectiveTargetsPool(objective, pool, binding.Namespace); err != nil {
 			return desiredBindingState{}, newStateError(conditionTypeInvalidRef, "InvalidObjectiveRef", err.Error())
 		}
 	}
@@ -493,7 +495,7 @@ func (r *InferenceIdentityBindingReconciler) computeDesiredState(
 		"rendered identity from inference intent",
 		logKeyMode, identity.Mode,
 		logKeySpiffeID, identity.SpiffeID,
-		logKeyClusterSPIFFEID, identity.Name,
+		logKeyClusterSPIFFEID, desiredClusterSPIFFEID(binding, identity).GetName(),
 		logKeySelectors, identity.Selectors,
 		logKeyPodSelector, identity.PodSelector,
 		logKeyObjective, identity.ObjectiveRef,
