+++ internal/controller/inferenceidentitybinding_gvk.go
@@ -3,13 +3,13 @@ package controller
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/sonda-red/kleym/internal/identity"
+	"github.com/sonda-red/kleym/internal/gaie"
 )
 
 func (r *InferenceIdentityBindingReconciler) resolveObjectiveGVKs() []schema.GroupVersionKind {
-	return identity.ResolveObjectiveGVKs(r.availableObjectiveGVKs)
+	return gaie.ResolveObjectiveGVKs(r.availableObjectiveGVKs)
 }
 
 func (r *InferenceIdentityBindingReconciler) resolvePoolGVKs() []schema.GroupVersionKind {
-	return identity.ResolvePoolGVKs(r.availablePoolGVKs)
+	return gaie.ResolvePoolGVKs(r.availablePoolGVKs)
 }
