+++ internal/controller/inferenceidentitybinding_logging_test.go
@@ -116,7 +116,6 @@ func TestReconcileClusterSPIFFEIDsLogsApplyDecisions(t *testing.T) {
 	scheme := newCollisionTestScheme(t)
 	binding := newPoolOnlyBinding("binding-log-apply", "objective-a")
 	identity := renderedIdentity{
-		Name:     "desired-clusterspiffeid",
 		Mode:     kleymv1alpha1.InferenceIdentityBindingModePoolOnly,
 		SpiffeID: "spiffe://kleym.sonda.red/ns/default/pool/pool-a",
 		Selectors: []string{
@@ -132,11 +131,11 @@ func TestReconcileClusterSPIFFEIDsLogsApplyDecisions(t *testing.T) {
 	}
 
 	drifted := desiredClusterSPIFFEID(binding, identity)
+	expectedName := drifted.GetName()
 	drifted.Object["spec"] = map[string]any{"spiffeIDTemplate": "spiffe://wrong"}
 
-	staleIdentity := identity
-	staleIdentity.Name = "stale-clusterspiffeid"
-	stale := desiredClusterSPIFFEID(binding, staleIdentity)
+	stale := desiredClusterSPIFFEID(binding, identity)
+	stale.SetName("stale-clusterspiffeid")
 
 	reconciler := &InferenceIdentityBindingReconciler{
 		Client: fake.NewClientBuilder().
@@ -151,7 +150,7 @@ func TestReconcileClusterSPIFFEIDsLogsApplyDecisions(t *testing.T) {
 	}
 
 	logs.requireEntry(t, "updating drifted managed ClusterSPIFFEID", map[string]string{
-		logKeyClusterSPIFFEID: "desired-clusterspiffeid",
+		logKeyClusterSPIFFEID: expectedName,
 		logKeySpiffeID:        identity.SpiffeID,
 	})
 	logs.requireEntry(t, "deleting stale managed ClusterSPIFFEID", map[string]string{
