+++ internal/gaie/gvk_test.go
@@ -1,4 +1,4 @@
-package identity
+package gaie
 
 import (
 	"reflect"
